Minor style fixes, and remove an unneeded build dependency.
clen

+++ src/base/vof.f90
@@ -133,12 +133,12 @@ module vof
      ! Interface to C function cs_vof_source_term
 
      subroutine vof_drift_term(imrgra, nswrgp, imligp, iwarnp, epsrgp,    &
-                                   climgp, pvar  , pvara , smbrs )            &
+                               climgp, pvar  , pvara , smbrs )            &
        bind(C, name='cs_vof_drift_term')
        use, intrinsic :: iso_c_binding
        implicit none
-       integer, intent(in) :: imrgra, imligp, iwarnp, nswrgp
-       double precision, intent(in) :: epsrgp, climgp
+       integer(c_int), intent(in) :: imrgra, imligp, iwarnp, nswrgp
+       real(kind=c_double), intent(in) :: epsrgp, climgp
        real(kind=c_double), dimension(*), intent(in) :: pvar, pvara
        real(kind=c_double), dimension(*), intent(inout) :: smbrs
      end subroutine vof_drift_term
